requirements.txt
celery
requests
psycopg2-binary
redis
python-dotenv